alsa-mixer: Add an profile for EVO4 interface
The Audient EVO4 USB audio interface supports loopback audio, so the
desktop audio set out to it can also be included back into the input
audio. In some situations this is not what you want (e.g. when you use
it for conference calling).

This change introduces and additional profile which disables loopback
audio so the user can choose whether loopback audio is enabled or not.