virtual sink: Factor out common code
This patch moves the code for the virtual sink callbacks and initialization to
a separate file. The code is re-factored, extended and built as library, so that
it can be used by other virtual sinks as well. The suspend-virtual-on-master-suspend
fix for the ladspa sink (!68) was incorporated into the common code as well as the
sink part of !78 which fixes a crash with stacked virtual sinks. !68 has a bug
which leaves a virtual sink unavailable when the master sink disappears. This bug
is also fixed.

Additionally, fixed block size filters, fixed window size filters  and updating filter
parameters are supported by the library. Fixed window size filters always deliver the
same number of frames to the filter, padding new data with history frames if necessary.
Rewinding can be disabled or or limited to the number of frames that the virtual sink
supports. Thanks to Alexander E. Patrakov for pointing out how to rewind fixed block
size filters.

The library implements following command line arguments if they are enabled in
valid_modargs: sink_name, sink_properties, sink_input_properties, force_flat_volume,
remix, resample_method and autoloaded.