bluetooth: Switch codecs without tearing down sink/source/thread
The main source of routing audio to another sink on codec switch is the
PA sink disappearing, together with its IO thread and other primitives.
This is annoying for users but also wasteful; all we need to do for a
codec switch is make the sink temporarily unavailable, release the
transport, perform the usual `SetConfiguration` on another `SEP` and
`Acquire` the resulting stream that is made available, to finally
reconfigure the sink with the new sampling rate and format and make the
sink available again.