source: Remove output_from_master field
The only source still using the output_from_master field of the source structure
was module-echo-cancel. Due to the differences between the echo-cancel source
and the other virtual sources, consolidation did not seem useful. To faciliate
removing the output_from_master field from the source, a dummy vsource structure
was created for module-echo-cancel. After that, the output_from_master field was
removed.