alsa-mixer: add multichannel profile for Tascam Model12
Add new profile for multichannel input and output with the *Tascam
Model12* USB interface [1]. The 12 input and 10 output channels aren't
detected natively with the default profile.

The stereo profile maps channel 1 and 2 to `front-left`/`front-right`. The
interface allows to switch the main stereo downmix to channels 1 and 2 via
the built-in menu.

In the multichannel profile, input channels 1 to 10 are mapped as `aux0`
to `aux9`, respectively. The main stereo downmix is then mapped to
`front-left`/`front-right` as a more sensible default. Similarly, the output
channels 1 to 8 are mapped as `aux0` to `aux7`, respectively. Output channels
9/10 are then mapped to `front-left`/`front-right` as they are a stereo pair
and therefore better suited for stereo output than channels 1 and 2.

1: https://www.tascam.eu/en/model12