client-conf: Check XDG_SESSION_TYPE
If systemd is in use, the XDG_SESSION_TYPE will probably be set (usually
to "wayland", "x11", or "tty").  If XDG_SESSION_TYPE is set to something
other than "wayland", connect to the X11 server to request properties.

Commit 79a8efb45cb8304967baa2e8c6d6156478ce19d9 stopped connecting to
the X server unless SSH_CONNECTION is set in the environment, to avoid
launching a useless xwayland process. However, there are many more
circumstances in which a remote X server may be in use. For example,
XDMCP or VNC.

This does not completely fix #1412, but it does improve the situation.