bluetooth: Add no volume control quirk for Thinkplus XT99
When this headset works in HFP/HSP profile, the input/output volume
adjustment doesn't work if the headset is playing sth or capturing
sth. Through the btmon log, we could see "+VGS" and "+VGM" were sent
to headset and headset replied "command success", but the volume was
not changed.

It is something like once the SCO connecton is built, the volume
couldn't be changed by "+VGS" and "+VGM" commands. Here add a quirk
for this headset, skip to set set_sink_volume and set_source_volume,
after this change, the pulseaudio will apply the software volume
adjustment to this headset in the HFP/HSP mode. This change doesn't
impact A2DP volume adjustment for this headset.

Signed-off-by: Hui Wang <hui.wang@canonical.com>