echo-cancel: Support old webrtc-audio-processing library
PulseAudio requires webrtc-audio-processing-1 >= 1.0 since v17.0, but
Debian hasn't managed to package that version yet. The dependency is
only used for the echo-cancel module, and the API exposed by the webrtc
component doesn't appear to be changed.

It looks like we can revert the changes to allow building with an older
version, but it would be better to have the code support both versions
to make it easier to transition from early library versions to later
versions. As a step towards that, check in an older copy of the webrtc
component from commit b16b107171f2, and build with that when we have the
old library version.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>